requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
PyGithub==1.59.1
gitpython==3.1.40
requests==2.31.0
pydantic==2.4.2
google-generativeai==0.3.1
python-multipart==0.0.6
beautifulsoup4==4.12.2
lxml==4.9.3